リザルトの遷移問題解消？

+++ resources/views/games/result_guest.blade.php
@@ -56,7 +56,7 @@
 let isAutoRedirect = false;
 
 window.addEventListener('beforeunload', (event) => {
-    if (!isAutoRedirect) {
+    if (!isAutoRedirect && window.location.href !== '/home/{{ $room->id }}') {
         fetch(`{{ route('removeGameRoomGuest', ['room' => $room->id]) }}`, {
             method: 'DELETE',
             headers: {
